requirements.txt
asgiref==3.3.4
asn1crypto==1.4.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
azure-core==1.14.0
azure-storage-blob==12.8.1
boto3==1.17.83
botocore==1.20.83
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
ci-info==0.2.0
click==8.0.0
configobj==5.0.6
configparser==5.0.2
cryptography==3.4.7
decorator==4.4.2
Django==3.2.3
etelemetry==0.2.2
filelock==3.0.12
future==0.18.2
greenlet==1.1.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.3.0
isodate==0.6.0
jmespath==0.10.0
lxml==4.6.3
msrest==0.6.21
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.0
numpy==1.20.3
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.9
pandas==1.2.4
pathlib==1.0.1
Pillow==8.2.0
prov==2.0.0
pycparser==2.20
pycryptodomex==3.10.1
pydot==1.4.2
PyJWT==2.1.0
PyMuPDF==1.18.13
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyxnat==1.4
rdflib==5.0.0
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.4.2
scipy==1.6.3
simplejson==3.17.2
six==1.16.0
snowflake-connector-python==2.4.3
snowflake-sqlalchemy==1.2.4
SQLAlchemy==1.4.15
sqlparse==0.4.1
traits==6.2.0
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
